Backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
pdfplumber
python-docx
sentence-transformers
scikit-learn